Revert "Merge branch 'develop' into release/2.5.0"
This reverts commit 54b25a8124a3d459678d91aae836f73789d194e1, reversing
changes made to 59959b85bb99d1922bfbca229cfd927c9a3ca778.

+++ cdap-app-fabric/src/main/java/co/cask/cdap/test/RuntimeStats.java
@@ -33,7 +33,7 @@ public final class RuntimeStats {
 
   private static ConcurrentMap<String, AtomicLong> counters = Maps.newConcurrentMap();
 
-  public static void count(String name, long count) {
+  public static void count(String name, int count) {
     AtomicLong oldValue = counters.putIfAbsent(name, new AtomicLong(count));
     if (oldValue != null) {
       oldValue.addAndGet(count);
