+++ cdap-watchdog/src/main/java/co/cask/cdap/metrics/data/Interpolators.java
@@ -39,7 +39,7 @@ public final class Interpolators {
     }
 
     @Override
-    public long interpolate(TimeValue start, TimeValue end, long ts) {
+    public int interpolate(TimeValue start, TimeValue end, long ts) {
       Preconditions.checkNotNull(start);
       Preconditions.checkNotNull(end);
       Preconditions.checkArgument((ts <= end.getTime()) && (ts >= start.getTime()));
@@ -55,7 +55,7 @@ public final class Interpolators {
       return maxAllowedGap;
     }
 
-    protected abstract long limitedInterpolate(TimeValue start, TimeValue end, long ts);
+    protected abstract int limitedInterpolate(TimeValue start, TimeValue end, long ts);
   }
 
   /**
@@ -74,7 +74,7 @@ public final class Interpolators {
     }
 
     @Override
-    protected long limitedInterpolate(TimeValue start, TimeValue end, long ts) {
+    protected int limitedInterpolate(TimeValue start, TimeValue end, long ts) {
       return (ts < end.getTime()) ? start.getValue() : end.getValue();
     }
   }
@@ -94,11 +94,11 @@ public final class Interpolators {
     }
 
     @Override
-    protected long limitedInterpolate(TimeValue start, TimeValue end, long ts) {
+    protected int limitedInterpolate(TimeValue start, TimeValue end, long ts) {
       long deltaX = ts - start.getTime();
       long totalX = end.getTime() - start.getTime();
-      long totalY = end.getValue() - start.getValue();
-      long deltaY = (int) (totalY * deltaX / totalX);
+      int totalY = end.getValue() - start.getValue();
+      int deltaY = (int) (totalY * deltaX / totalX);
       return start.getValue() + deltaY;
     }
   }
